Added test to check if libnlopt exists. Testing nc_galaxy_acf.
There is a bug in nlopt package, when built with c++ support
it builds only the libnlopt_cxx library which cannot be linked
to a pure C library. Also, in the nlopt.pc it informs that the
library is linked using -lnlopt, even though the correct would
be -lnlopt_cxx. Thus, we added a extra check to see if -lnlopt
is available.