Added a internal version of Cuba. Fixed minor typos and updated autogen to use autoreconf.
It is no longer necessary to have an installed version of Cuba. If
it is not available in the system the library compiles its own internal
version of Cuba. The same was done recently for levmar.
Updated README.md to explain these points.

Removed cubature lib support (now it always use Cuba).

Added option for different inclusing of redshift errors in NcSNIADistCov.

Removed the call for cholesky_decomp in cholesky_inverse, now the user must
call both.