Added new diagnostics to NcmMSetCatalog, max ESS and Heidelberger and Welch's convergence diagnostic.
Both can be applied to any NcmStatsVec object or through the NcmMSetCatalog interface.
In the latter the test can be applied to individual chains, to the full catalog and to the ensemble
average. Added an option to NcmFitESMCMC to automatically trim  the catalog during an ESMCMC run
using the diagnostics to estimate the best burnin. Added options to run the diagnostics to mcat_analize.